ESSIFI-167: simplified the regex, moved test cases to proper place, added more test cases
